tests: Move remote realm billing tests into test_stripe_remote_realm.
test_stripe had grown to 10,000 lines, mixing Cloud realm billing
flows with the flows for self-hosted realms and servers.  Move the
remote realm classes into their own module, named to keep the
"stripe" marker for modules that use @mock_stripe fixtures, and to
avoid confusion with test_remote_billing, which covers the
authentication flow rather than billing.

Fixtures are keyed by test function name, so none need regenerating;
the new module does need listing for full-suite fixture regeneration.
